LPS-107005 Move reindex logic from SLA as well
